requirements.txt
Flask>=1.1.2
python-jose>=3.2.0
requests>=2.24.0
